# Sistema_Crud_API_PIBIC/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: SistemaAPI PIBIC
  version: 1.0.0
  description: |
    API REST para gerenciamento e resolucao de casos clinicos academicos.
    As operacoes de IA aceitam somente dados sinteticos ou previamente
    desidentificados e toda saida exige revisao humana antes da publicacao.
    O recurso e educacional e nao substitui decisao clinica.
  contact:
    name: Equipe SistemaAPI PIBIC
  license:
    name: Uso academico
...
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /casos/{id}/ia/gerar:
    post:
      summary: Gera conteudo clinico com apoio da IA
      description: |
        Completa os campos nao informados pelo professor e salva o conteudo
        gerado no caso clinico. Aceita somente dados sinteticos ou previamente
        desidentificados. A saida deve ser revisada por um professor.
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CasoClinicoIARequest"
      responses:
        "200":
          description: Conteudo clinico gerado e salvo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasoClinicoIAResponse"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "409":
          $ref: "#/components/responses/Conflito"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
        "429":
          $ref: "#/components/responses/LimiteAtingido"
        "502":
          $ref: "#/components/responses/FalhaProvedorIa"
        "503":
          $ref: "#/components/responses/CapacidadeIaIndisponivel"
        "504":
          $ref: "#/components/responses/TempoEsgotadoIa"
  /casos/{id}/ia/ajustar:
    post:
      summary: Ajusta conteudo clinico existente com apoio da IA
      description: |
        Reescreve o conteudo clinico mais recente do caso conforme o tipo de
        ajuste solicitado. Aceita somente dados sinteticos ou previamente
        desidentificados. A saida deve ser revisada por um professor.
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CasoClinicoIAAjusteRequest"
      responses:
        "200":
          description: Conteudo clinico ajustado e salvo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasoClinicoIAResponse"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "409":
          $ref: "#/components/responses/Conflito"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
        "429":
          $ref: "#/components/responses/LimiteAtingido"
        "502":
          $ref: "#/components/responses/FalhaProvedorIa"
        "503":
          $ref: "#/components/responses/CapacidadeIaIndisponivel"
        "504":
          $ref: "#/components/responses/TempoEsgotadoIa"
  /casos/{id}/ia/perguntas/gerar:
    post:
      summary: Gera perguntas para um caso clinico com apoio da IA
      description: |
        Gera e persiste perguntas para um caso com conteudo clinico criado
        manualmente ou por IA. Aceita somente dados sinteticos ou previamente
        desidentificados. Perguntas e gabaritos exigem revisao humana.
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GerarPerguntasIaRequestDTO"
      responses:
        "201":
          description: Perguntas geradas e persistidas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerguntaResponse"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/Conflito"
        "413":
          $ref: "#/components/responses/CorpoRequisicaoExcedido"
        "429":
          $ref: "#/components/responses/LimiteAtingido"
        "502":
          $ref: "#/components/responses/FalhaProvedorIa"
        "503":
          $ref: "#/components/responses/CapacidadeIaIndisponivel"
        "504":
          $ref: "#/components/responses/TempoEsgotadoIa"
  /casos/{id}/ia/auditoria:
    get:
      summary: Lista a trilha de auditoria das geracoes por IA do caso
      description: |
        Retorna somente metadados, hashes e referencias de proveniencia; o
        conteudo clinico e a saida do modelo nao sao duplicados nesta trilha.
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
      responses:
        "200":
          description: Pagina de eventos de geracao por IA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginaAuditoriaGeracaoIa"
        "403":
          $ref: "#/components/responses/AcessoNegado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: UUID estável para repetir com segurança a mesma tentativa de IA. Durante a retenção, uma chave não pode ser reutilizada com outro caso, operação ou payload. O ledger retém apenas metadados e IDs de resultado; a retenção padrão é de uma hora.
      schema:
        type: string
        format: uuid
        maxLength: 64
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    page:
      name: page
      in: query
      description: Indice da pagina, iniciando em zero.
      schema:
        type: integer
        minimum: 0
        default: 0
    size:
      name: size
      in: query
      description: Quantidade de elementos por pagina.
      schema:
        type: integer
        minimum: 1
        default: 20
    sort:
      name: sort
      in: query
      description: Ordenacao no formato campo,direcao.
      schema:
        type: string
        example: nome,asc
  responses:
    CorpoRequisicaoExcedido:
      description: O corpo da requisicao excede o limite configurado.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: urn:sistema-api-pibic:problem:corpo-requisicao-excedido
            title: Corpo da requisicao excedido
            status: 413
            detail: O corpo da requisicao excede o limite permitido
            instance: /auth/login
            timestamp: "2026-08-22T20:00:00Z"
            correlationId: 7c973582-36bc-45a7-980f-285901008ad8
            erro: O corpo da requisicao excede o limite permitido
    LimiteAtingido:
      description: Limite de tentativas, cota de IA ou concorrencia atingido.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
        Retry-After:
          description: Segundos que o cliente deve aguardar antes de tentar novamente.
          schema:
            type: integer
            format: int64
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: urn:sistema-api-pibic:problem:limite-uso-ia
            title: Limite de uso da IA
            status: 429
            detail: O limite de geracoes de IA por minuto foi atingido
            instance: /casos/42/ia/gerar
            timestamp: "2026-08-22T20:00:00Z"
            correlationId: 7c973582-36bc-45a7-980f-285901008ad8
            erro: O limite de geracoes de IA por minuto foi atingido
    CapacidadeIaIndisponivel:
      description: |
        A integracao de IA nao foi configurada ou nenhum provedor possui
        capacidade disponivel no momento. Quando presente, Retry-After informa
        os segundos ate a proxima tentativa recomendada.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
        Retry-After:
          description: Segundos que o cliente deve aguardar antes de tentar novamente.
          schema:
            type: integer
            format: int64
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            configuracaoAusente:
              value:
                type: urn:sistema-api-pibic:problem:servico-indisponivel
                title: Servico indisponivel
                status: 503
                detail: Configure a integracao de IA antes de usar este recurso
                instance: /casos/42/ia/gerar
                timestamp: "2026-08-22T20:00:00Z"
                correlationId: 7c973582-36bc-45a7-980f-285901008ad8
                erro: Configure a variavel IA_CHAVE_API antes de usar a IA
            capacidadeEsgotada:
              value:
                type: urn:sistema-api-pibic:problem:capacidade-ia-esgotada
                title: Capacidade de IA esgotada
                status: 503
                detail: A capacidade gratuita dos provedores de IA foi esgotada temporariamente
                instance: /casos/42/ia/gerar
                timestamp: "2026-08-22T20:00:00Z"
                correlationId: 7c973582-36bc-45a7-980f-285901008ad8
                erro: A capacidade gratuita dos provedores de IA foi esgotada temporariamente
    TempoEsgotadoIa:
      description: O provedor de IA excedeu o tempo limite configurado.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: urn:sistema-api-pibic:problem:tempo-esgotado-ia
            title: Tempo esgotado no provedor de IA
            status: 504
            detail: O provedor de IA excedeu o tempo limite da requisicao
            instance: /casos/42/ia/gerar
            timestamp: "2026-08-22T20:00:00Z"
            correlationId: 7c973582-36bc-45a7-980f-285901008ad8
            erro: O provedor de IA excedeu o tempo limite da requisicao
    FalhaProvedorIa:
      description: O provedor de IA rejeitou a solicitacao ou devolveu uma resposta invalida.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    DadosInvalidos:
      description: A requisicao possui campos, JSON, parametros ou tipos invalidos.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    AcessoNegado:
      description: O usuario autenticado nao possui permissao para a operacao.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NaoEncontrado:
      description: O recurso solicitado nao foi encontrado.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflito:
      description: O estado atual ou a concorrencia dos dados impede a operacao.
      headers:
        X-Correlation-Id:
          $ref: "#/components/headers/CorrelationId"
        Retry-After:
          description: Segundos ate a repeticao recomendada quando o conflito for transitorio.
          schema:
            type: integer
            format: int64
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  headers:
    CorrelationId:
      description: Identificador da requisicao para rastreamento ponta a ponta.
      schema:
        type: string
        minLength: 1
        maxLength: 64
  schemas:
...
    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - instance
        - timestamp
        - correlationId
        - erro
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        timestamp:
          type: string
          format: date-time
        correlationId:
          type: string
        erro:
          type: string
          deprecated: true
          description: Alias legado de detail; mantido temporariamente para compatibilidade.
        errors:
          type: object
          additionalProperties:
            type: string
          description: Erros por campo quando a validacao falha.
        campos:
          type: object
          deprecated: true
          additionalProperties:
            type: string
          description: Alias legado de errors; mantido temporariamente para compatibilidade.
...
    Paciente:
      type: object
      required:
        - idCaso
        - nome
        - profissao
        - sexo
        - idade
        - estadoCivil
        - altura
        - peso
      properties:
        idPaciente:
          type: integer
          format: int64
          readOnly: true
        idCaso:
          type: integer
          format: int64
          minimum: 1
        nome:
          type: string
          maxLength: 150
        profissao:
          type: string
          maxLength: 120
        sexo:
          type: string
          enum: [MASCULINO, FEMININO, OUTRO, NAO_INFORMADO]
        idade:
          type: integer
          minimum: 0
          maximum: 130
        estadoCivil:
          type: string
          enum: [SOLTEIRO, CASADO, DIVORCIADO, VIUVO, SEPARADO, UNIAO_ESTAVEL, NAO_INFORMADO]
        altura:
          type: string
          maxLength: 20
        peso:
          type: string
          maxLength: 20
    ConteudoClinico:
      type: object
      required:
        - idCaso
        - sintomas
        - contexto
        - examClinico
        - antecClinico
        - diagEsperado
      properties:
        idConteudo:
          type: integer
          format: int64
          readOnly: true
        idCaso:
          type: integer
          format: int64
          minimum: 1
        sintomas:
          type: string
          maxLength: 10000
        contexto:
          type: string
          maxLength: 10000
        examClinico:
          type: string
          maxLength: 10000
        antecClinico:
          type: string
          maxLength: 10000
        diagEsperado:
          type: string
          maxLength: 10000
    CasoClinicoRequest:
      type: object
      required:
        - titulo
        - disciplina
        - areaSaude
        - estilo
        - especialidade
        - nivelDificuldade
      properties:
        idProfessor:
          type: integer
          format: int64
          nullable: true
        titulo:
          type: string
          maxLength: 200
        disciplina:
          type: string
          maxLength: 120
        areaSaude:
          type: string
          maxLength: 120
        estilo:
          type: string
          maxLength: 60
        especialidade:
          type: string
          maxLength: 120
        objetivoAprendizagem:
          type: string
          maxLength: 10000
        nivelDificuldade:
          type: string
          enum: [BAIXA, MEDIA, ALTA]
        tempoLimiteMinutos:
          type: integer
          minimum: 15
          maximum: 480
          default: 60
          description: Tempo total para responder o caso completo.
    CasoClinicoResponse:
      allOf:
        - $ref: "#/components/schemas/CasoClinicoRequest"
        - type: object
          properties:
            idCaso:
              type: integer
              format: int64
            status:
              type: string
              enum: [RASCUNHO, PUBLICADO, ARQUIVADO]
            dataCriacao:
              type: string
              format: date-time
            dataAtualizacao:
              type: string
              format: date-time
    CasoClinicoIARequest:
      type: object
      description: Campos preenchidos parcialmente pelo professor; os campos ausentes serao completados pela IA.
      required:
        - dadosSinteticosOuDesidentificados
      properties:
        sintomas:
          type: string
          maxLength: 10000
        contexto:
          type: string
          maxLength: 10000
        examClinico:
          type: string
          maxLength: 10000
        antecClinico:
          type: string
          maxLength: 10000
        diagEsperado:
          type: string
          maxLength: 10000
        permitirComplementoIa:
          type: boolean
          description: Permite que a IA complemente dados ausentes do paciente.
        incluirResultadosExamesClinicos:
          type: boolean
          description: Permite incluir resultados de exames laboratoriais, imagem ou complementares.
        informacoesAdicionaisPaciente:
          type: string
          maxLength: 10000
        dadosSinteticosOuDesidentificados:
          type: boolean
          enum: [true]
          description: Confirmacao obrigatoria de que nenhum dado pessoal identificavel sera enviado ao provedor.
    CasoClinicoIAAjusteRequest:
      type: object
      required:
        - tipoAjuste
        - dadosSinteticosOuDesidentificados
      properties:
        tipoAjuste:
          type: string
          enum: [REGERAR, SIMPLIFICAR, COMPLEXIFICAR, PERSONALIZADO]
        instrucao:
          type: string
          maxLength: 2000
        dadosSinteticosOuDesidentificados:
          type: boolean
          enum: [true]
          description: Confirmacao obrigatoria de que o contexto enviado esta sintetico ou desidentificado.
    CasoClinicoIAResponse:
      type: object
      properties:
        idConteudo:
          type: integer
          format: int64
        idCaso:
          type: integer
          format: int64
        sintomas:
          type: string
          maxLength: 10000
        contexto:
          type: string
          maxLength: 10000
        examClinico:
          type: string
          maxLength: 10000
        antecClinico:
          type: string
          maxLength: 10000
        diagEsperado:
          type: string
          maxLength: 10000
        completo:
          $ref: "#/components/schemas/CasoClinicoCompleto"
    CasoClinicoCompleto:
      type: object
      properties:
        caso:
          $ref: "#/components/schemas/CasoClinicoResponse"
        pacientes:
          type: array
          items:
            $ref: "#/components/schemas/Paciente"
        conteudosClinicos:
          type: array
          items:
            $ref: "#/components/schemas/ConteudoClinico"
        perguntas:
          type: array
          items:
            $ref: "#/components/schemas/PerguntaResponse"
    AuditoriaGeracaoIa:
      type: object
      required:
        - id
        - idCaso
        - idUsuario
        - operacao
        - provedor
        - modelo
        - versaoPrompt
        - hashContexto
        - hashSaida
        - referenciaResultado
        - quantidadeItens
        - dadosDesidentificadosConfirmados
        - dataGeracao
      properties:
        id:
          type: integer
          format: int64
        idCaso:
          type: integer
          format: int64
        idUsuario:
          type: integer
          format: int64
        operacao:
          type: string
          enum: [GERAR_CASO, AJUSTAR_CASO, GERAR_PERGUNTAS]
        provedor:
          type: string
        modelo:
          type: string
        versaoPrompt:
          type: string
        hashContexto:
          type: string
          description: Hash SHA-256 do contexto desidentificado enviado.
        hashSaida:
          type: string
          description: Hash SHA-256 da saida recebida, sem duplicar seu conteudo.
        referenciaResultado:
          type: string
        quantidadeItens:
          type: integer
          format: int32
          minimum: 1
        dadosDesidentificadosConfirmados:
          type: boolean
        duracaoProvedorMs:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        tokensEntrada:
          type: integer
          format: int32
          minimum: 0
          nullable: true
        tokensSaida:
          type: integer
          format: int32
          minimum: 0
          nullable: true
        correlationId:
          type: string
          maxLength: 64
          nullable: true
        dataGeracao:
          type: string
          format: date-time
    GerarPerguntasIaRequestDTO:
      type: object
      description: Configuracao da geracao de perguntas; a confirmacao de desidentificacao e obrigatoria.
      required:
        - dadosSinteticosOuDesidentificados
      properties:
        quantidade:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
          example: 5
        tipo:
          type: string
          enum: [MULTIPLA_ESCOLHA, DISCURSIVA, VERDADEIRO_FALSO, DIAGNOSTICO, CONDUTA_CLINICA]
          default: MULTIPLA_ESCOLHA
          example: MULTIPLA_ESCOLHA
        quantidadeAlternativas:
          type: integer
          minimum: 2
          maximum: 5
          default: 4
          example: 4
        instrucoesAdicionais:
          type: string
          maxLength: 2000
          example: Priorize raciocinio diagnostico e conduta inicial.
        dadosSinteticosOuDesidentificados:
          type: boolean
          enum: [true]
          description: Confirmacao obrigatoria de que o contexto enviado esta sintetico ou desidentificado.
    AlternativaPergunta:
      type: object
      properties:
        id:
          type: integer
          format: int64
        letra:
          type: string
          maxLength: 10
        texto:
          type: string
          maxLength: 10000
        correta:
          type: boolean
...
    PerguntaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        idCaso:
          type: integer
          format: int64
        texto:
          type: string
        tipo:
          type: string
          enum: [MULTIPLA_ESCOLHA, DISCURSIVA, VERDADEIRO_FALSO, DIAGNOSTICO, CONDUTA_CLINICA]
          example: MULTIPLA_ESCOLHA
        gabarito:
          type: string
        resposta:
          type: string
        alternativas:
          type: array
          items:
            $ref: "#/components/schemas/AlternativaPergunta"
...
    PaginaBase:
      type: object
      required:
        - totalElements
        - totalPages
        - number
        - size
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
...
    PaginaAuditoriaGeracaoIa:
      allOf:
        - $ref: "#/components/schemas/PaginaBase"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/AuditoriaGeracaoIa"
...